requirements.txt
requests==2.31.0
boto3==1.34.0
faiss-cpu==1.10.0
numpy==1.26.4
streamlit==1.32.0
plotly==5.20.0
python-dotenv==1.0.1
protobuf<5
